Switch from `IncrementalSake` to era dependent `InstantStake`
* Simplify proposal deposit addition to pool stake

* Allow for credential with zero stake to be ommitted from the Stake map

* Make `calculatePoolDistr` account for lack of zeros in `Stake`

+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rewards.hs
@@ -369,4 +369,4 @@ mkPoolRewardInfo
         guard (hk == ppId pool)
         fromCompact <$> VMap.lookup (KeyHashObj o) (unStake stake)
       Coin ostake = Set.foldl' accOwnerStake mempty (ppOwners pool)
-      sigma = fromIntegral pstakeTot %? unCoin totalStake
+      sigma = pstakeTot %? unCoin totalStake
